move leader election to different file
